Add user_id to responses

+++ src/Models/Drawer.hs
@@ -22,17 +22,19 @@ import Data.Maybe (catMaybes)
 import Database.PostgreSQL.Simple.ToField (ToField(toField))
 
 data Drawer = Drawer
-  { dId        :: Integer
+  { dId            :: Integer
+  , dUserId        :: UserId
   , dStorageUnitId :: Integer
-  , dLevel     :: Integer
-  , dNote      :: String } deriving Generic
+  , dLevel         :: Integer
+  , dNote          :: String } deriving Generic
 
 instance FromRow Drawer where
-  fromRow = Drawer <$> field <*> field <*> field <*> field
+  fromRow = Drawer <$> field <*> field <*> field <*> field <*> field
 
 instance ToJSON Drawer where
-  toEncoding (Drawer id' storageUnitId level note) =
+  toEncoding (Drawer id' userId storageUnitId level note) =
     pairs $    "id"              .= id'
+            <> "user_id"         .= userId
             <> "storage_unit_id" .= storageUnitId
             <> "level"           .= level
             <> "note"            .= note
@@ -42,17 +44,17 @@ instance ToJSON Drawer where
 getAllDrawers :: UserId -> Integer -> IO [Drawer]
 getAllDrawers userId paramStorageUnitId = do
   withConn $ \conn -> query conn queryString (userId, paramStorageUnitId)
-  where queryString = "SELECT id, storage_unit_id, level, note FROM drawers WHERE user_id = ? AND storage_unit_id = ?;"
+  where queryString = "SELECT id, user_id, storage_unit_id, level, note FROM drawers WHERE user_id = ? AND storage_unit_id = ?;"
 
 getDrawer :: UserId -> Integer -> IO (Maybe Drawer)
 getDrawer userId paramId = do
   withConn $ \conn -> query conn queryString (paramId, userId) >>= resultsToMaybeDrawer
-  where queryString = "SELECT id, storage_unit_id, level, note FROM drawers WHERE id = ? AND user_id = ? LIMIT 1"
+  where queryString = "SELECT id, user_id, storage_unit_id, level, note FROM drawers WHERE id = ? AND user_id = ? LIMIT 1"
 
 createDrawer :: UserId -> Integer -> Integer -> String -> IO (Maybe Drawer)
 createDrawer userId paramStorageUnitId paramLevel paramNote = do
   withConn $ \conn -> query conn queryString (userId, paramStorageUnitId, userId, paramLevel, paramNote) >>= resultsToMaybeDrawer
-  where queryString = "INSERT INTO drawers (user_id, storage_unit_id, level, note) VALUES (?, (SELECT id FROM storage_units where id = ? AND user_id = ?), ?, ?) RETURNING id, storage_unit_id, level, note"
+  where queryString = "INSERT INTO drawers (user_id, storage_unit_id, level, note) VALUES (?, (SELECT id FROM storage_units where id = ? AND user_id = ?), ?, ?) RETURNING id, user_id, storage_unit_id, level, note"
 
 updateDrawer :: UserId -> Integer -> Maybe Integer -> Maybe Integer -> Maybe String -> IO (Maybe Drawer)
 updateDrawer userId paramId paramStorageUnitId paramLevel paramNote = do
@@ -68,8 +70,8 @@ updateDrawer userId paramId paramStorageUnitId paramLevel paramNote = do
           , toField <$> Just paramId
           , toField <$> Just userId ]
         updatesString = if L.null updateList then mempty else mconcat $ L.intersperse ", " updateList
-        updateQueryString = "UPDATE drawers SET " <> updatesString <> " WHERE id = ? AND user_id = ? RETURNING id, storage_unit_id, level, note"
-        selectQueryString = "SELECT id, storage_unit_id, level, note FROM drawers WHERE id = ? AND user_id = ? LIMIT 1"
+        updateQueryString = "UPDATE drawers SET " <> updatesString <> " WHERE id = ? AND user_id = ? RETURNING id, user_id, storage_unit_id, level, note"
+        selectQueryString = "SELECT id, user_id, storage_unit_id, level, note FROM drawers WHERE id = ? AND user_id = ? LIMIT 1"
 
     resultsToMaybeDrawer =<< if L.null updateList
       then query conn selectQueryString (paramId, userId)
@@ -78,14 +80,15 @@ updateDrawer userId paramId paramStorageUnitId paramLevel paramNote = do
 deleteDrawer :: UserId -> Integer -> IO (Maybe Drawer)
 deleteDrawer userId paramId = do
   withConn $ \conn -> query conn queryString (paramId, userId) >>= resultsToMaybeDrawer
-  where queryString = "DELETE FROM drawers WHERE id = ? AND user_id = ? RETURNING id, storage_unit_id, level, note"
+  where queryString = "DELETE FROM drawers WHERE id = ? AND user_id = ? RETURNING id, user_id, storage_unit_id, level, note"
 
 -- helper functions
 
-resultsToMaybeDrawer :: [(Integer, Integer, Integer, String)] -> IO (Maybe Drawer)
+resultsToMaybeDrawer :: [(Integer, UserId, Integer, Integer, String)] -> IO (Maybe Drawer)
 resultsToMaybeDrawer = \case
-  [(resId, resStorageUnitId, resLevel, resNote)] -> return $ Just $ Drawer { dId = resId
-                                                                           , dStorageUnitId = resStorageUnitId
-                                                                           , dLevel = resLevel
-                                                                           , dNote = resNote }
+  [(resId, resUserId, resStorageUnitId, resLevel, resNote)] -> return $ Just $ Drawer { dId = resId
+                                                                                      , dUserId = resUserId
+                                                                                      , dStorageUnitId = resStorageUnitId
+                                                                                      , dLevel = resLevel
+                                                                                      , dNote = resNote }
   _ -> return Nothing
